refactor: move. danger modal parts to shared admin
